section and subsection done

+++ src/components/CategoryPage/CategoryTable.vue
@@ -217,9 +217,36 @@ export default {
   computed: mapGetters(["allCategories",'allSubSections']),
   methods: {
     deleteCategory(id) {
-      this.$store.dispatch("deleteCategory", id)
-      console.log(this.x+"hi");
-      // this.$router.go()
+      this.$swal
+        .fire({
+          title: " هل انت متاكد ؟ ",
+          text: " لن تستطيع استرجاع هذا القسم مجددا ! ",
+          icon: "warning",
+          showCancelButton: true,
+          confirmButtonColor: "#1C64F2",
+          cancelButtonColor: "#d33",
+          confirmButtonText: " نعم, انا متاكد ",
+          cancelButtonText: "الغاء",
+        })
+        .then((result) => {
+          if (result.isConfirmed) {
+            this.$store.dispatch("deleteCategory", id);
+            this.$swal.fire({
+              icon: "success",
+              title: " تم ",
+              text: " تم حذف القسم بنجاح ",
+              confirmButtonColor: "#16a34a",
+            });
+          } else {
+            this.$swal.fire({
+              icon: "error",
+              title: " الغاء ",
+              text: " تم الغاء عملية الحذف ",
+              confirmButtonColor: "#16a34a",
+              confirmButtonText: "حسنا",
+            });
+          }
+        });
     },
       
     ...mapActions(['fetchCategories'])
